backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
aiosqlite
ffmpeg-python
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
python-multipart
